requirements.txt
# Vietnamese SLU Data Augmentation Requirements
# ==============================================

# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.64.0

# Text processing
regex>=2022.7.9
unidecode>=1.3.4

# JSON processing
jsonlines>=3.1.0

# CLI and utilities
click>=8.1.0
colorama>=0.4.5

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Optional: For advanced text processing
# transformers>=4.20.0  # Uncomment if using Vietnamese language models
# torch>=1.12.0         # Uncomment if using Vietnamese language models